backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic>=2.9.0
pydantic-settings>=2.6.0
supabase==2.3.0
python-dotenv==1.0.0
pillow>=10.3.0
pytesseract==0.3.10
trafilatura==1.6.3
beautifulsoup4==4.12.2
httpx>=0.24.0,<0.25.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
openai>=1.12.0
